refactor(admin): extract block-form field-renderers module
Verbatim move of the field HTML builders (imageFieldHtml, fileFieldHtml,
primitiveInputHtml, renderPrimitiveField, renderArrayPrimitiveItem,
renderArrayObjectItem, renderArrayField) and their single-consumer icon
constants (trash, chevrons, drag handle, file picker) into
block-form/field-renderers.ts. Only renderPrimitiveField and
renderArrayField are exported — they are all mountBlockForm calls; the
rest stay module-internal.

Part of #38, slice 4/6.
